refactor: pickStore 폴더 제거
