+++ frontend/techpick/src/stores/pickStore.ts
