+++ frontend/techpick/src/stores/pickStore.type.ts
